// src/app/models/filiais.model.ts
export interface Filiais{
    id?: any;
    razao_social: string;
    nome_fantasia: string;
    cnpj: string;
    inscricao_estadual: string;
    logradouro: string;
    cep: string;
    bairro: string;
    cidade: string;
    uf: string;
    ibge: string;
    codigo_pais: string;
    pais: string;
    ativo: string;
    complemento: string;
    numero: string;
    email: string;
    telefone: string;
    celular: string;
    contato: string;
    tipo_empresa: string;
    porte: string;
    data_abertura: Date;
    natureza_juridica: string;
    tipo: string;
}